requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
beautifulsoup4==4.12.3
branca==0.7.2
certifi==2024.7.4
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap4==2024.1
cryptography==43.0.1
cssselect2==0.7.0
defusedxml==0.8.0rc2
Django==5.0.7
django-admin-adminlte==1.0.5
django-cors-headers==4.6.0
django-crispy-forms==2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
drf-yasg==1.21.8
et-xmlfile==1.1.0
folium==0.17.0
gunicorn==21.2.0
html5lib==1.1
idna==3.7
inflection==0.5.1
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==2.1.5
mysqlclient==2.2.6
numpy==2.0.1
oauthlib==3.2.2
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pycparser==2.22
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.10.1
pypdf==4.3.1
pypng==0.20220715.0
pyserial==3.5
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.2
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.5
sqlparse==0.5.1
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.2.2
webencodings==0.5.1
whitenoise==6.5.0
xhtml2pdf==0.2.16
xyzservices==2024.6.0
